Change default base_url to nil
Change the default argument for base_rule in login/3 to nil in order to properly make use of the short circuit logic in the function.

As it is now someone calling this function has  to explicitly supply nil if they have already configured their base_url elsewhere (e.g. mix.exs).